Settings overhaul
