+++ apps/ezplayer-ui-electron/src/router/router.tsx
@@ -4,22 +4,30 @@ import { Navigate, RouteObject } from 'react-router';
 import { SuspenseLoader } from '@ezplayer/shared-ui-components';
 
 import {
+    AudioSettings,
     CloudPage,
     CreateEditPlaylist,
     JukeboxScreen,
     JukeboxFullScreen,
+    JukeboxSettings,
+    PlayerCloudRegistrationDialog,
     PlayerScreen,
+    PlayerSettings,
     PlaylistList,
     Routes as ROUTES,
     Schedule,
     SidebarLayout,
     SongList,
+    ShowFolderSettings,
     ShowStatusScreen,
-    PlaybackSettingsDrawer,
+    SettingsDrawer,
     Preview3DPage,
+    UISettings,
+    ViewerSettings,
     toRouteChildren,
 } from '@ezplayer/player-ui-components';
-import type { MenuRoute } from '@ezplayer/player-ui-components';
+import type { MenuRoute, SettingsSection } from '@ezplayer/player-ui-components';
+import { useState } from 'react';
 
 import TableChartTwoToneIcon from '@mui/icons-material/TableChartTwoTone';
 import PlayArrow from '@mui/icons-material/PlayArrow';
@@ -29,6 +37,11 @@ import MusicIcon from '@mui/icons-material/MusicNoteTwoTone';
 import ViewInArIcon from '@mui/icons-material/ViewInAr';
 import DisplaySettingsIcon from '@mui/icons-material/DisplaySettings';
 import CloudIcon from '@mui/icons-material/Cloud';
+import FolderIcon from '@mui/icons-material/Folder';
+import ContrastIcon from '@mui/icons-material/Contrast';
+import LanguageIcon from '@mui/icons-material/Language';
+import VolumeUpIcon from '@mui/icons-material/VolumeUp';
+import TuneIcon from '@mui/icons-material/Tune';
 
 import { AddSongDialogElectron } from '../components/song/AddSongDialogElectron';
 import { WelcomeScreen, WELCOME_ROUTE } from '../modules/Welcome/WelcomeScreen';
@@ -47,6 +60,62 @@ const ErrorPage = Loader(lazy(() => ERROR_PAGE));
 
 //const getStatusArea = ()=>[<ConnectivityStatus key="connectivity-status" />];
 const getStatusArea = () => [];
+
+const ElectronSettingsPage = () => {
+    const [cloudOpen, setCloudOpen] = useState(false);
+    const sections: SettingsSection[] = [
+        {
+            key: 'showFolder',
+            label: 'Show Folder',
+            icon: <FolderIcon sx={{ fontSize: 56 }} />,
+            content: <ShowFolderSettings />,
+        },
+        {
+            key: 'ui',
+            label: 'UI',
+            icon: <ContrastIcon sx={{ fontSize: 56 }} />,
+            content: <UISettings />,
+        },
+        {
+            key: 'viewer',
+            label: 'Viewer',
+            icon: <LanguageIcon sx={{ fontSize: 56 }} />,
+            title: 'Viewer Control',
+            content: <ViewerSettings />,
+        },
+        {
+            key: 'jukebox',
+            label: 'Jukebox',
+            icon: <PlayArrow sx={{ fontSize: 56 }} />,
+            content: <JukeboxSettings />,
+        },
+        {
+            key: 'audio',
+            label: 'Audio',
+            icon: <VolumeUpIcon sx={{ fontSize: 56 }} />,
+            content: <AudioSettings />,
+        },
+        {
+            key: 'cloud',
+            label: 'Cloud',
+            icon: <CloudIcon sx={{ fontSize: 56 }} />,
+            onClick: () => setCloudOpen(true),
+        },
+        {
+            key: 'player',
+            label: 'Player',
+            icon: <TuneIcon sx={{ fontSize: 56 }} />,
+            title: 'Player Settings',
+            content: <PlayerSettings />,
+        },
+    ];
+    return (
+        <>
+            <SettingsDrawer title="Settings" statusArea={getStatusArea()} sections={sections} />
+            <PlayerCloudRegistrationDialog open={cloudOpen} onClose={() => setCloudOpen(false)} />
+        </>
+    );
+};
 
 const menuRoutes: MenuRoute[] = [
     {
@@ -91,7 +160,7 @@ const menuRoutes: MenuRoute[] = [
     },
     {
         path: ROUTES.PLAYBACKSETTINGS,
-        element: <PlaybackSettingsDrawer title="Settings" statusArea={getStatusArea()} />,
+        element: <ElectronSettingsPage />,
         sidebar: { icon: <DisplaySettingsIcon />, label: 'Settings' },
     },
     {
