+++ packages/player-ui-components/package.json
@@ -1,7 +1,7 @@
 {
     "name": "@ezplayer/player-ui-components",
     "private": true,
-    "version": "0.12.8",
+    "version": "0.13.2",
     "type": "module",
     "main": "./dist/index.js",
     "types": "./dist/index.d.ts",
