+++ packages/player-ui-components/src/index.ts
@@ -19,7 +19,14 @@ export { PlayerScreen } from './components/player/PlayerScreen';
 export { JukeboxArea, JukeboxFullScreen, JukeboxScreen } from './components/jukebox/JukeboxScreen';
 export { ShowStatusScreen } from './components/status/ShowStatusScreen';
 export { StatsDialog } from './components/status/StatsDialog';
-export { PlaybackSettingsDrawer } from './components/playback-settings/PlaybackSettingsDrawer';
+export { SettingsDrawer } from './components/playback-settings/SettingsDrawer';
+export type { SettingsSection } from './components/playback-settings/SettingsDrawer';
+export { ShowFolderSettings } from './components/playback-settings/sections/ShowFolderSettings';
+export { UISettings } from './components/playback-settings/sections/UISettings';
+export { ViewerSettings } from './components/playback-settings/sections/ViewerSettings';
+export { JukeboxSettings } from './components/playback-settings/sections/JukeboxSettings';
+export { AudioSettings } from './components/playback-settings/sections/AudioSettings';
+export { PlayerSettings } from './components/playback-settings/sections/PlayerSettings';
 export { PlayerCloudRegistrationDialog } from './components/player-cloud-registration/PlayerCloudRegistrationDialog';
 export { CloudPage } from './components/cloud/CloudPage';
 export { ColorPaletteDialog } from './components/theme/ColorPaletteDialog';
